requirements.txt
selenium
webdriver-manager
pytest-html